requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
filelock==3.6.0
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
numpy==1.21.5
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
six==1.16.0
tokenizers==0.11.6
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
transformers==4.18.0
typing_extensions==4.1.1
urllib3==1.26.9
zipp==3.8.0
